Update README storage example to use managed disks
Apparently, even with the appropriate permissions on the role, k8s can't
seem to properly create unmanaged disks.  Unsure if this is a limitation
of k8s that goes against the docs or if it's some aspect of the types of
VMs created by Juju, but the upshot is that we have to use managed
disks.